+++ samples/src/main/java/com/example/DetokenizeExample.java
@@ -42,7 +42,7 @@ public class DetokenizeExample {
             ResponseToken res = null;
             try {
                 String filePath = "<your_credentials_file_path>";
-                res = Token.GenerateToken(filePath);
+                res = Token.generateBearerToken(filePath);
             } catch (SkyflowException e) {
                 e.printStackTrace();
             }