+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [1.2.0] - 2022-01-04
+## [1.2.0] - 2022-01-11
 
 ### Added
 - Logging functionality